+++ external/spdlog/CMakeLists.txt
@@ -8,4 +8,6 @@ opendaq_dependency(
     GIT_REPOSITORY      https://github.com/gabime/spdlog.git
     GIT_REF             v1.13.0
     EXPECT_TARGET       spdlog::spdlog
+    PATCH_FILES
+        ${CMAKE_CURRENT_SOURCE_DIR}/patches/001-periodic_worker_init_func.patch
 )
